feat: added support for srt files
