+++ requirements.txt
@@ -1,6 +1,31 @@
+appdirs==1.4.4
+black==21.5b2
+certifi==2020.12.5
+chardet==4.0.0
+click==7.1.2
+distlib==0.3.1
 docopt==0.6.2
+filelock==3.0.12
 fpdf2==2.2.0
+greenlet==1.1.1
+mypy-extensions==0.4.3
 numpy==1.19.5
 opencv-python==4.5.1.48
+pathspec==0.8.1
 Pillow==8.1.0
+pipenv==2020.11.15
+playwright==1.15.0
+pyee==8.2.2
+pysrt==1.1.2
+python-dotenv==0.19.0
+regex==2020.11.13
+rope==0.18.0
+six==1.15.0
+srt==3.5.0
+toml==0.10.2
+typed-ast==1.4.2
+typing-extensions==3.7.4.3
+virtualenv==20.4.4
+virtualenv-clone==0.5.4
+websockets==10.0
 webvtt-py==0.4.6
